2nd wrap vaporizes asteroids + start of wave counter
Currently working through a bug in the wave counter. If you clear 1
wave and die, it will say that you’ve cleared 3 waves on the game over
screen.

To test this:
I have the fragments the spawn upon destroying a large asteroid
disabled.

The requirement of wrapping the shot around the screen is disabled as
well.